translation/requirements.txt
fastapi
uvicorn[standard]
lxml
pandas
pycountry
wget
fasttext-wheel
transformers
nltk
html5lib
